requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20

# LLM Providers
openai>=1.0.0
anthropic>=0.8.0
huggingface_hub>=0.19.0

# Vector Stores & Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document Loaders & Processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
unstructured>=0.11.0

# Web & API Tools
requests>=2.31.0
aiohttp>=3.9.0
google-search-results>=2.4.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tiktoken>=0.5.2

# Development & Jupyter
ipykernel>=6.27.0
jupyter>=1.0.0
notebook>=7.0.0

# Optional but recommended
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0